(#301) Windows now builds with warning level 4 (previously level 3) and treats...

Closes #301

See merge request hdtn-v4/hdtn!291

+++ common/config/include/InductsConfig.h
@@ -29,7 +29,7 @@
 #include <utility>
 #include <tuple>
 #include "JsonSerializable.h"
-#include <boost/filesystem.hpp>
+#include <boost/filesystem/path.hpp>
 #include "config_lib_export.h"
 
 struct induct_element_config_t {
@@ -73,7 +73,7 @@ struct induct_element_config_t {
     uint64_t remoteNodeId;
 
     //specific to stcp and tcpcl
-    uint32_t keepAliveIntervalSeconds;
+    uint16_t keepAliveIntervalSeconds;
 
     //specific to tcpcl version 3 (servers)
     uint64_t tcpclV3MyMaxTxSegmentSizeBytes;
